A neuron is evaluate only once for a network step
Classe neuron now contains a date and an output variable.
When a network does a step, it gives its own timestamp to each neuron. If the neuron is up-to-date, it returns its output (getOutput), otherwise it evaluates (evaluate).
This is very useful for combinational network. Layer1 will evaluate, then Layer2, etc... Without this implementation, each neuron of Layer2 will call the evaluate function of each neuron of Layer1, and so time for evaluate a whole network becomes exponantial.